requirements.txt
# Core ML libraries
scikit-learn>=1.0.0
xgboost>=1.5.0
lightgbm>=3.3.0

# Deep learning




# Data processing
pandas>=1.3.0
numpy>=1.21.0

# Utilities
pyyaml>=6.0
joblib>=1.1.0
tqdm>=4.62.0

# Class imbalance handling
imbalanced-learn>=0.8.0

# Visualization (optional)
matplotlib>=3.5.0
seaborn>=0.11.0
